Move ObjectHeapData to heap/object.rs
